requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
altgraph @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
astunparse==1.6.3
attrs==25.4.0
audioread==3.1.0
basic-pitch==0.4.0
cachetools==6.2.4
cattrs==25.3.0
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.8
cloudpickle==3.1.2
contourpy==1.3.0
coremltools==9.0
cycler==0.12.1
decorator==5.2.1
demucs==4.0.1
dora_search==0.1.12
einops==0.8.1
exceptiongroup==1.3.1
fastapi==0.128.0
filelock==3.19.1
flatbuffers==25.12.19
fonttools==4.60.2
fsspec==2025.10.0
future @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
gast==0.7.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.14.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
jsonpickle==4.1.1
julius==0.2.7
keras==2.15.0
kiwisolver==1.4.7
lameenc==1.8.1
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.43.0
macholib @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
mido==1.3.3
mir_eval==0.8.2
ml-dtypes==0.3.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
music21==8.3.0
namex==0.1.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
oauthlib==3.3.1
omegaconf==2.3.0
openunmix==1.3.0
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
pretty_midi==0.2.11
protobuf==4.25.8
pyaml==25.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
resampy==0.4.2
retrying==1.4.2
rich==14.2.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
six @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
soundfile==0.13.1
soxr==1.0.0
starlette==0.49.3
submitit==1.5.4
sympy==1.14.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
torch==2.8.0
torchaudio==2.8.0
tqdm==4.67.1
treetable==0.2.6
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.39.0
webcolors==24.11.1
Werkzeug==3.1.4
wrapt==1.14.2
zipp==3.23.0
